Quantum/master

+++ src/user/user.controller.js
@@ -1,5 +1,5 @@
 const ErrorHandler = require("../../utils/errorHandler");
-const { userModel, otpModel } = require("./user.model");
+const { userModel, otpModel, verifiedModel } = require("./user.model");
 const catchAsyncError = require("../../utils/catchAsyncError");
 const sendEmail = require("../../utils/sendEmail");
 const generateOTP = require("../../utils/otpGenerator");
@@ -9,7 +9,13 @@ const { Op } = require("sequelize");
 const { db } = require("../../config/database");
 const { eventModel } = require("../events/event.model");
 const { notificationModel } = require("../notification");
+// const { createStripeCustomers } = require("../../utils/stripe");
+const secret_key = process.env.STRIPE_SECRET_KEY;
+const stripe = require("stripe")(secret_key);
+const { firebase } = require("../../utils/firebase");
+const bcrypt = require("bcryptjs");
 
+const messaging = firebase.messaging();
 const getMsg = (otp) => {
   return `<html lang="en">
   <head>
@@ -60,53 +66,86 @@ const createNotification = async (userId, text, title, userAvatar) => {
   console.log("Notification created successfully");
 };
 
-exports.register = catchAsyncError(async (req, res, next) => {
-  console.log("register", req.body);
-  const { email, dob, mobile_no } = req.body;
-  const imageFile = req.file;
-  const imageUrl = imageFile && (await s3Uploadv2(imageFile));
-  let user;
-  const prevUser = await userModel.findOne({ email: email });
+exports.sendDummyToken = catchAsyncError(async (req, res, next) => {
+  const { firebaseToken } = req.body;
+  const message = {
+    notification: {
+      title: "Hello from FCM!",
+      body: "This is a test message from FCM.",
+    },
+    token: null,
+  };
+  messaging
+    .send(message)
+    .then((response) => {
+      res.status(200).send(response);
+      console.log("fcm errorrrrr", response);
+    })
+    .catch((error) => {
+      console.log("fcm errorrrrr", error);
 
-  if (prevUser && !prevUser.isVerified) {
-    await prevUser.update({ ...req.body });
-    user = prevUser;
-    console.log(prevUser);
-  } else {
-    user = imageUrl
-      ? await userModel.create({
-          ...req.body,
-          role: "User",
-          dob: new Date(dob),
-          avatar: imageUrl.Location,
-        })
-      : await userModel.create({
-          ...req.body,
-          role: "User",
-          dob: new Date(dob),
-        });
-  }
-
-  const otp = generateOTP();
-
-  await storeOTP({ otp, userId: user.id });
-
-  try {
-    const message = getMsg(otp);
-    await sendEmail({
-      email: user.email,
-      subject: "Verify Registration OTP",
-      message,
+      res.status(400).send(error);
     });
-    res
-      .status(StatusCodes.CREATED)
-      .json({ message: `OTP sent to ${user.email} successfully` });
-  } catch (error) {
-    return next(
-      new ErrorHandler(error.message, StatusCodes.INTERNAL_SERVER_ERROR)
-    );
-  }
 });
+
+// exports.register = catchAsyncError(async (req, res, next) => {
+//   console.log("register", req.body);
+//   const { email, username, dob, fireBaseToken, mobile_no } = req.body;
+//   const imageFile = req.file;
+//   const imageUrl = imageFile && (await s3Uploadv2(imageFile));
+//   let user;
+//   const prevUser = await userModel.findOne({ email: email });
+
+//   if (!fireBaseToken) {
+//     return next(
+//       new ErrorHandler("Fire base token is required", StatusCodes.NOT_FOUND)
+//     );
+//   }
+
+//   const stripeCustomerId = await createStripeCustomer(email, username);
+
+//   if (prevUser && !prevUser.isVerified) {
+//     await prevUser.update({ ...req.body });
+//     user = prevUser;
+//     console.log(prevUser);
+//   } else {
+//     user = imageUrl
+//       ? await userModel.create({
+//         ...req.body,
+//         role: "User",
+//         fcm_token: fireBaseToken,
+//         dob: new Date(dob),
+//         avatar: imageUrl.Location,
+//       })
+//       : await userModel.create({
+//         ...req.body,
+//         role: "User",
+//         fcm_token: fireBaseToken,
+//         customerId: stripeCustomerId,
+//         dob: new Date(dob),
+//       });
+//   }
+
+//   const otp = generateOTP();
+
+//   await storeOTP({ otp, userId: user.id });
+
+//   try {
+//     const message = getMsg(otp);
+//     await sendEmail({
+//       email: user.email,
+//       subject: "Verify Registration OTP",
+//       message,
+//     });
+//     res
+//       .status(StatusCodes.CREATED)
+//       .json({ message: `OTP sent to ${user.email} successfully` });
+//   } catch (error) {
+//     return next(
+//       new ErrorHandler(error.message, StatusCodes.INTERNAL_SERVER_ERROR)
+//     );
+//   }
+// });
 
 // exports.verifyRegisterOTP = catchAsyncError(async (req, res, next) => {
 //   const { otp, email } = req.body;
@@ -148,14 +187,91 @@ exports.register = catchAsyncError(async (req, res, next) => {
 //   const token = user.getJWTToken();
 //   res.status(StatusCodes.CREATED).json({ success: true, user, token });
 // });
+
+exports.register = catchAsyncError(async (req, res, next) => {
+  console.log("register", req.body);
+  const { email, dob, fireBaseToken } = req.body;
+  const imageFile = req.file;
+  const imageUrl = imageFile && (await s3Uploadv2(imageFile));
+  let user;
+  const prevUser = await verifiedModel.findOne({ email: email });
+
+  if (!fireBaseToken) {
+    return next(
+      new ErrorHandler("Fire base token is required", StatusCodes.NOT_FOUND)
+    );
+  }
+
+  if (prevUser && !prevUser.isVerified) {
+    await prevUser.update({ ...req.body });
+    user = prevUser;
+    console.log(prevUser);
+  } else {
+    user = imageUrl
+      ? await verifiedModel.create({
+        ...req.body,
+        role: "User",
+        fcm_token: fireBaseToken,
+        dob: new Date(dob),
+        avatar: imageUrl.Location,
+      })
+      : await verifiedModel.create({
+        ...req.body,
+        role: "User",
+        fcm_token: fireBaseToken,
+        dob: new Date(dob),
+      });
+  }
+
+  const otp = generateOTP();
+
+  await storeOTP({ otp, userId: user.id });
+
+  try {
+    const message = getMsg(otp);
+    await sendEmail({
+      email: user.email,
+      subject: "Verify Registration OTP",
+      message,
+    });
+    res
+      .status(StatusCodes.CREATED)
+      .json({ message: `OTP sent to ${user.email} successfully` });
+  } catch (error) {
+    return next(
+      new ErrorHandler(error.message, StatusCodes.INTERNAL_SERVER_ERROR)
+    );
+  }
+});
+
 exports.verifyRegisterOTP = catchAsyncError(async (req, res, next) => {
   const { otp, email } = req.body;
   if (!otp || !email) {
     return next(new ErrorHandler("Missing OTP", StatusCodes.BAD_REQUEST));
   }
 
-  const user = await userModel.findOne({ where: { email } });
-  if (!user) {
+  const userVerified = await verifiedModel.findOne({
+    where: { email },
+    attributes: [
+      "id",
+      "email",
+      "password",
+      "username",
+      "mobile_no",
+      "country",
+      "dob",
+      "isVerified",
+      "role",
+      "gender",
+      "avatar",
+      "fcm_token",
+      "customerId",
+      "createdAt",
+      "updatedAt",
+    ],
+  });
+
+  if (!userVerified) {
     return next(
       new ErrorHandler(
         "User not found. Please check the entered email.",
@@ -165,7 +281,6 @@ exports.verifyRegisterOTP = catchAsyncError(async (req, res, next) => {
   }
 
   const otpInstance = await otpModel.findOne({ where: { otp } });
-  console.log("otpInstance", otpInstance);
 
   if (!otpInstance) {
     return next(
@@ -182,29 +297,70 @@ exports.verifyRegisterOTP = catchAsyncError(async (req, res, next) => {
   if (!otpDuration) {
     // OTP found but expired
     await otpModel.destroy({ where: { id: otpInstance.id } });
-    await userModel.destroy({ where: { email } });
+    // await userModel.destroy({ where: { email } });
     return next(
       new ErrorHandler("OTP has been expired.", StatusCodes.BAD_REQUEST)
     );
   }
 
   // Proceed with user verification
-  user.isVerified = true;
-  await user.save();
+  userVerified.isVerified = true;
+  await userVerified.save();
+
+  const user = await userModel.create({
+    id: userVerified.id.toString(),
+    email: userVerified.email,
+    password: userVerified.password,
+    username: userVerified.username,
+    mobile_no: userVerified.mobile_no,
+    country: userVerified.country,
+    dob: userVerified.dob,
+    isVerified: userVerified.isVerified,
+    role: userVerified.role,
+    gender: userVerified.gender,
+    avatar: userVerified.avatar,
+    fcm_token: userVerified.fcm_token,
+    createdAt: userVerified.createdAt,
+    updatedAt: userVerified.updatedAt,
+    deletedAt: null,
+  });
+
+  const stripeCustomer = await stripe.customers.create({
+    email: user.email,
+    name: user.username,
+  });
+
+  if (!stripeCustomer) {
+    return next(new ErrorHandler("Stripe error", StatusCodes.BAD_REQUEST));
+  }
+
+  let customer = {};
+
+  if (stripeCustomer) customer.customerId = stripeCustomer.id;
+
+  await user.update(customer);
 
   // Destroy the OTP instance after successful verification
   await otpModel.destroy({ where: { id: otpInstance.id } });
+  // await verifiedModel.destroy({ where: { id: user.id } });
 
   // Generate JWT token for the user
   const token = user.getJWTToken();
+  const notificationText = `Welcome to Lyvechat and unlock the events in the modern day of virtuality`;
+  await createNotification(user.id, notificationText, "Account Created", user.avatar);
 
-  // Respond with success and user details along with token
   res.status(StatusCodes.CREATED).json({ success: true, user, token });
 });
 
 exports.login = catchAsyncError(async (req, res, next) => {
   console.log("login", req.body);
-  const { email, password } = req.body;
+  const { email, password, fireBaseToken } = req.body;
+
+  if (!fireBaseToken) {
+    return next(
+      new ErrorHandler("Fire base token is required", StatusCodes.NOT_FOUND)
+    );
+  }
 
   const user = await userModel
     .scope("withPassword")
@@ -234,6 +390,12 @@ exports.login = catchAsyncError(async (req, res, next) => {
   }
 
   const token = user.getJWTToken();
+
+  if (fireBaseToken) {
+    user.fcm_token = fireBaseToken;
+    await user.save();
+  }
+
   res.status(StatusCodes.OK).json({ user, token });
 });
 
@@ -244,7 +406,7 @@ exports.resendOTP = catchAsyncError(async (req, res, next) => {
     return next(new ErrorHandler("Please enter your email.", 400));
   }
 
-  const user = await userModel.findOne({ where: { email } });
+  const user = await verifiedModel.findOne({ where: { email } });
   if (!user) {
     return next(
       new ErrorHandler("Please register or User doesn't exist.", 400)
@@ -278,6 +440,7 @@ exports.forgotPassword = catchAsyncError(async (req, res, next) => {
   }
 
   const user = await userModel.findOne({ where: { email: req.body.email } });
+  console.log("usss", user);
 
   if (!user) {
     return next(
@@ -315,24 +478,47 @@ exports.forgotPassword = catchAsyncError(async (req, res, next) => {
 
 exports.verifyOtp = catchAsyncError(async (req, res, next) => {
   const { otp } = req.body;
+
   if (!otp) {
     return next(new ErrorHandler("Missing OTP", StatusCodes.BAD_REQUEST));
   }
 
   const otpInstance = await otpModel.findOne({ where: { otp } });
 
-  if (!otpInstance || otpInstance.isValid()) {
-    if (otpInstance) {
-      await otpModel.destroy({ where: { id: otpInstance.id } });
-    }
-
+  if (!otpInstance) {
+    await otpModel.destroy({ where: { id: otpInstance.id } });
     return next(
       new ErrorHandler(
-        "OTP is invalid or has been expired",
+        "Invalid OTP. Please check the entered OTP.",
         StatusCodes.BAD_REQUEST
       )
     );
   }
+
+  const otpDuration = await otpInstance.isValid();
+
+  // Check OTP validity and expiration
+  if (!otpDuration) {
+    // OTP found but expired
+    await otpModel.destroy({ where: { id: otpInstance.id } });
+    // await userModel.destroy({ where: { email } });
+    return next(
+      new ErrorHandler("OTP has been expired.", StatusCodes.BAD_REQUEST)
+    );
+  }
+
+  // if (!otpInstance || otpInstance.isValid()) {
+  //   if (otpInstance) {
+  //     await otpModel.destroy({ where: { id: otpInstance.id } });
+  //   }
+
+  //   return next(
+  //     new ErrorHandler(
+  //       "OTP is invalid or has been expired",
+  //       StatusCodes.BAD_REQUEST
+  //     )
+  //   );
+  // }
 
   await otpModel.destroy({ where: { id: otpInstance.id } });
 
@@ -608,7 +794,7 @@ exports.followCreator = catchAsyncError(async (req, res, next) => {
 
   if (userId === creatorId) {
     return next(
-      new ErrorHandler("you can not follow yourself", StatusCodes.FORBIDDEN)
+      new ErrorHandler("You cannot follow yourself", StatusCodes.FORBIDDEN)
     );
   }
 
@@ -635,6 +821,29 @@ exports.followCreator = catchAsyncError(async (req, res, next) => {
     "follow",
     currUser.avatar
   );
+
+  let token = targetCreator.fcm_token;
+  console.log("Target user FCM token:", token);
+
+  const fcmMessage = {
+    notification: {
+      title: "New Follower",
+      body: `${currUser.username} started following you.`,
+    },
+    token,
+    data: {
+      type: "follow",
+    },
+  };
+
+  try {
+    await messaging.send(fcmMessage);
+    console.log("Push notification sent successfully.");
+  } catch (error) {
+    // Handle error if FCM token is expired or invalid
+    console.error("Error sending push notification:", error);
+    // Log the error and proceed with the follow operation
+  }
 
   res
     .status(StatusCodes.CREATED)
@@ -726,7 +935,7 @@ exports.getCreatorFollowers = catchAsyncError(async (req, res, next) => {
 
   const followers = await userModel.findAll({
     ...query,
-    attributes: ["id", "username", "avatar"],
+    attributes: ["id", "username", "avatar", "createdAt"],
   });
 
   // Get the following users for the current user
@@ -797,7 +1006,7 @@ exports.getCreatorFollowing = catchAsyncError(async (req, res, next) => {
 
   const followings = await userModel.findAll({
     ...query,
-    attributes: ["id", "username", "avatar"],
+    attributes: ["id", "username", "avatar", "createdAt"],
   });
 
   // Get the following users for the current user
